% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cat_ls}
\alias{cat_ls}
\title{Example categorical raster (fine_dat)}
\format{
A raster layer object.
}
\source{
Sciaini M, Fritsch M, Scherer C, Simpkins CE. NLMR and landscapetools: An integrated environment
    for simulating and modifying neutral landscape models in R. Methods in Ecology and Evolution. 2018;
    00:1-9. https://doi.org/10.1111/2041-210X.13076

Marco Sciaini and Matthias Fritsch (2018). landscapetools: Landscape Utility Toolbox. R package version 0.4.0.
    https://CRAN.R-project.org/package=landscapetools
}
\usage{
cat_ls
}
\description{
An example map to show functionality on categorical surfaces.
}
\details{
Generated with \code{nlm_mpd()} from \code{NLMR} and classified with \code{util_classify()} from \code{landscapetools}.
}
\keyword{datasets}
