% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cont_ls}
\alias{cont_ls}
\title{Example continuous raster (fine_dat)}
\format{A raster layer object.}
\source{
Sciaini M, Fritsch M, Scherer C, Simpkins CE. NLMR and landscapetools: An integrated environment
    for simulating and modifying neutral landscape models in R. Methods in Ecology and Evolution. 2018;
    00:1-9. https://doi.org/10.1111/2041-210X.13076
}
\usage{
cont_ls
}
\description{
An example map to show functionality on continuous surfaces.
}
\details{
Generated with \code{nlm_mpd()} from \code{NLMR}.
}
\keyword{datasets}
