% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_torus.R
\name{create_torus}
\alias{create_torus}
\title{Pad a raster by a specified radius to create the effect of a torus. Allows for moving window analysis that avoids edge effects}
\usage{
create_torus(dat, r)
}
\arguments{
\item{dat}{The raster dataset to pad}

\item{r}{The radius by which to pad the raster}
}
\value{
raster. Original raster padded by radius r with torus effect
}
\description{
This function calculates a value for Shannon evenness for an input raster based on a list of landcover classes.
}
\keyword{raster}
\keyword{torus,}
