% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winmoveR_functions.R
\name{diversity}
\alias{diversity}
\title{Calculate Shannon evenness for a given neighbourhood and list of landcover types}
\usage{
diversity(dat, nbrhd, lc_class)
}
\arguments{
\item{dat}{The raster dataset on which to calculate the moving window Shannon evenness}

\item{nbrhd}{The neighbourhood on which to perform the calculation. Defined by raster::focalWeight()}

\item{lc_class}{The list of land cover classes for which Shannon evenness should be calculated.}
}
\value{
raster with Shannon evenness value for the neighbourhood surrounding each cell
}
\description{
This function calculates a value for Shannon evenness for each cell of a raster within a specified neighbourhood based on a list of landcover types.
}
\keyword{Shannon,}
\keyword{diversity,}
\keyword{evenness,}
\keyword{focal}
