% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_nbrhd.R
\name{diversity_nbrhd}
\alias{diversity_nbrhd}
\title{Calculate Shannon evenness for a given neighbourhood and list of landcover classes}
\usage{
diversity_nbrhd(dat, nbrhd, lc_class)
}
\arguments{
\item{dat}{The raster dataset on which to calculate the moving window Shannon evenness}

\item{nbrhd}{The neighbourhood on which to perform the calculation. Defined by raster::focalWeight(). If this is NULL, the calculation is performed on the full grid cell.}

\item{lc_class}{The list of land cover classes for which Shannon evenness should be calculated.}
}
\value{
raster with Shannon evenness value for the neighbourhood surrounding each cell
}
\description{
This function calculates a value for Shannon evenness for each cell of a raster within a specified neighbourhood based on a list of landcover classes.
}
\keyword{Shannon,}
\keyword{diversity,}
\keyword{evenness,}
\keyword{focal}
