% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g_sf}
\alias{g_sf}
\title{Example grid (coarse_dat)}
\format{
An sf object.
}
\source{
Sciaini M, Fritsch M, Scherer C, Simpkins CE. NLMR and landscapetools: An integrated environment
    for simulating and modifying neutral landscape models in R. Methods in Ecology and Evolution. 2018;
    00:1-9. https://doi.org/10.1111/2041-210X.13076
}
\usage{
g_sf
}
\description{
An example grid to show functionality when aggregating using an sf object.
}
\details{
Generated with \code{nlm_mpd()} and converted to sf.
}
\keyword{datasets}
