% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grainchanger-package.R
\docType{package}
\name{grainchanger-package}
\alias{grainchanger}
\alias{grainchanger-package}
\title{grainchanger: Moving-window and direct data aggregation}
\description{
Data aggregation via moving window or direct methods.
}
\author{
\strong{Maintainer}: Laura Graham \email{LauraJaneEGraham@gmail.com}

}
\keyword{internal}
