% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{mean}
\alias{mean}
\alias{mean.winmove}
\title{Arithmetic mean}
\usage{
mean(x, ...)

\method{mean}{winmove}(x, d, type, na.rm = FALSE, ...)
}
\arguments{
\item{x}{RasterLayer. The data over which to calculate the mean value within a moving
window}

\item{...}{further arguments passed to or from other methods}

\item{d}{numeric. If \code{type=circle}, the radius of the circle (in units of the
CRS). If \code{type=rectangle} the dimension of the rectangle (one or two numbers).
If \code{type=Gauss} the size of sigma, and optionally another number to determine
the size of the matrix returned (default is 3 times sigma)}

\item{type}{character. The shape of the moving window}

\item{na.rm}{logical. indicates whether \code{NA} values should be stripped before the
computation proceeds. \code{na.rm = FALSE} (default) is optimised to be much faster
than \code{na.rm = TRUE}. The latter should only be used if \code{NA} values are
present}
}
\value{
RasterLayer. A smoothed raster with the mean calculated within the specified
  moving window
}
\description{
An extension to \code{mean} for objects of class \code{winmove}
}
\examples{

# load required data
data(cont_ls)

# convert data to object of class winmove
cont_ls <- new("winmove", cont_ls)

# aggregate using a circular window with radius 3
d <- mean(cont_ls, d = 3, type = "circle")
}
\keyword{focal}
\keyword{mean}
\keyword{spatial}
