% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomove_agg.R
\name{nomove_agg}
\alias{nomove_agg}
\title{Direct data aggregation}
\usage{
nomove_agg(coarse_dat, fine_dat, agg_fun, is_grid = TRUE,
  quiet = FALSE, ...)
}
\arguments{
\item{coarse_dat}{sf, Raster* or Spatial* object. The coarse grain data
(response data) across which to calculate the aggregated function}

\item{fine_dat}{Raster* object. Raster* object. The fine grain data
(predictor / covariate data) to aggregate}

\item{agg_fun}{function The function to apply. The function fun should take
multiple numbers, and return a single number. For example mean, modal, min
or max. It should also accept a na.rm argument (or ignore it, e.g. as one
of the 'dots' arguments. For example, length will fail, but function(x,
...){na.omit(length(x))} works. See Details}

\item{is_grid}{logical. Use \code{TRUE} (default) if \code{g} contains only
rectangular cells (i.e. a grid). If \code{g} is any other polygon file,
this should be set to false}

\item{quiet}{logical. If \code{FALSE} (default) and \code{is_grid == TRUE}
the user gets a warning that the aggregation assumes all cells are
rectangular}

\item{...}{further arguments passed to or from other methods}
}
\value{
Raster (if input is Raster) or numeric vector (if input is sp or sf
  object) containing values calculated for each coarser cell
}
\description{
Calculate the value for a given function for each cell in a larger resolution
grid.
}
\examples{
# load required data
data(g_sf)
data(cont_ls)
data(cat_ls)

# aggregate using mean
d <- nomove_agg(g_sf, cont_ls, mean)

# aggregate using Shannon evenness
d <- nomove_agg(g_sf, cont_ls, shei, lc_class = 1:4)
}
\keyword{#'}
\keyword{(max}
\keyword{(scalar)}
\keyword{(vector}
\keyword{-}
\keyword{@details}
\keyword{Functions}
\keyword{Note}
\keyword{Proportion,}
\keyword{Range}
\keyword{Shannon}
\keyword{\code{future}}
\keyword{\code{grainchanger}}
\keyword{\code{lc_class}}
\keyword{\code{nomove_agg}}
\keyword{\code{plan(multiprocess)}}
\keyword{\code{prop}}
\keyword{\code{shdi}}
\keyword{\code{shei}}
\keyword{\code{var_range}}
\keyword{\item}
\keyword{\itemize{}
\keyword{additional}
\keyword{aggregate}
\keyword{are:}
\keyword{argument}
\keyword{be}
\keyword{built-in}
\keyword{can}
\keyword{currently}
\keyword{diversity,}
\keyword{evenness,}
\keyword{from}
\keyword{functions.}
\keyword{has}
\keyword{in}
\keyword{included}
\keyword{min)}
\keyword{or}
\keyword{package.}
\keyword{parallel}
\keyword{requires}
\keyword{run}
\keyword{scalar)}
\keyword{several}
\keyword{spatial,}
\keyword{that}
\keyword{the}
\keyword{using}
\keyword{}}
