% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomove_cell.R
\name{nomove_cell}
\alias{nomove_cell}
\title{Calculate the value of a given function for a raster within a particular cell of a larger grid}
\usage{
nomove_cell(cell, grid, dat, fn, ...)
}
\arguments{
\item{cell}{the row identifier of the grid cell}

\item{grid}{the grid from which to take the cell}

\item{dat}{The raster dataset on which to calculate the function}

\item{fn}{The function to apply}

\item{...}{further arguments passed to or from other methods}
}
\value{
numeric. The value of the function applied to the raster within the larger resolution cell
}
\description{
This function takes as input the row identifier of a cell within a large resolution grid and calculates the value for a given function for a smaller resolution raster dataset.
}
