% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomove_upscale.R
\name{nomove_upscale}
\alias{nomove_upscale}
\title{Upscale the function value}
\usage{
nomove_upscale(grid, dat, fn, ...)
}
\arguments{
\item{grid}{the grid across which to calculate the upscaled moving window function}

\item{dat}{The raster dataset on which to calculate the moving window function}

\item{fn}{The function to apply}

\item{...}{further arguments passed to or from other methods}
}
\value{
Numeric vector containing values calculated for each grid cell
}
\description{
Calculate the value for a given function for each cell in a larger resolution grid.
}
