% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poly_sf}
\alias{poly_sf}
\title{Example polygon (coarse_dat)}
\format{
An sf object.
}
\usage{
poly_sf
}
\description{
An example non-gridded coarse data to show functionality when aggregating using an sf object.
}
\details{
Generated with \code{sf::st_make_grid(sf::st_as_sfc(sf::st_bbox(cont_ls)), cellsize = 13, square = FALSE)}
}
\keyword{datasets}
