% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_nbrhd.R
\name{prop_nbrhd}
\alias{prop_nbrhd}
\title{Calculate proportion of a given land cover type within a given neighbourhood}
\usage{
prop_nbrhd(dat, nbrhd, lc_class)
}
\arguments{
\item{dat}{The raster dataset on which to calculate the land-cover proportion}

\item{nbrhd}{The neighbourhood on which to perform the calculation. Defined by raster::focalWeight(). If this is NULL, the calculation is performed on the full grid cell.}

\item{lc_class}{The list of land-cover classes for calculating the proportion}
}
\value{
raster with proportion of given land-cover classes in the neighbourhood surrounding each cell
}
\description{
This function calculates for each cell of a raster the proportion of a given land-cover type(s) within a specified neighbourhood.
}
\keyword{composition,}
\keyword{focal}
\keyword{proportion,}
