% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{var_range}
\alias{var_range}
\title{Variable range}
\usage{
var_range(dat, na.rm = TRUE)
}
\arguments{
\item{dat}{A raster or vector}
}
\value{
numeric. Range of the variable
}
\description{
This function calculates the range of a continuous variable
}
\keyword{focal}
\keyword{range,}
