% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winmove.R
\name{winmove}
\alias{winmove}
\title{Calculate the value of a given function for each cell, within a given neighbourhood, for a cell in a larger resolution grid.}
\usage{
winmove(dat, d, type, fun, ...)
}
\arguments{
\item{dat}{The raster dataset on which to calculate the moving window function}

\item{d}{numeric. If `type=circle`, the radius of the circle (in units of the CRS). If
`type=rectangle` the dimension of the rectangle (one or two numbers). If `type=Gauss` the
size of sigma, and optionally another number to determine the size of the matrix
returned (default is 3 times sigma)}

\item{type}{The shape of the moving window}

\item{fun}{The function to apply}

\item{...}{further arguments passed to or from other methods}

\item{cell}{the row identifier of the grid cell}

\item{grid}{the grid from which to take the cell}
}
\value{
A raster with the moving window values calculated
}
\description{
This function takes as input the row identifier of a cell within a large resolution grid and calculates the moving window function for a smaller resolution raster dataset using the `winmove_nbrhd()` function.
}
