% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winmoveR_functions.R
\name{winmove_cellmean}
\alias{winmove_cellmean}
\title{Calculate the mean value of an input raster}
\usage{
winmove_cellmean(winmove_cellr)
}
\arguments{
\item{winmove_cellr}{raster.}
}
\value{
numeric. The mean value of the input raster
}
\description{
This function takes a raster as input and calculates it's mean value. For my purposes, this is used to upscale.
}
