% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winmoveR_functions.R
\name{winmove_raster}
\alias{winmove_raster}
\title{Calculate the value of a given function for each cell, within a given neighbourhood.}
\usage{
winmove_raster(dat, radius, type = c("circle", "Gauss", "rectangle"), fn, ...)
}
\arguments{
\item{dat}{The raster dataset on which to calculate the moving window function}

\item{radius}{The radius of the moving window}

\item{type}{The shape of the moving window}

\item{fn}{The function to apply}

\item{...}{further arguments passed to or from other methods}
}
\value{
A raster with the moving window values calculated
}
\description{
This function takes as input a raster dataset, the radius and type of moving window (`circle`, `rectangle` or `Gauss`) and a function to apply across the moving window. A new raster with the moving window value for each cell is then output.
}
