% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winmove_upscale.R
\name{winmove_upscale}
\alias{winmove_upscale}
\title{Upscale the moving window value}
\usage{
winmove_upscale(grid, dat, radius, type, verbose = FALSE, fn, ...)
}
\arguments{
\item{grid}{the grid across which to calculate the upscaled moving window function}

\item{dat}{The raster dataset on which to calculate the moving window function}

\item{radius}{The radius of the moving window}

\item{type}{The shape of the moving window}

\item{fn}{The function to apply}

\item{...}{further arguments passed to or from other methods}
}
\value{
Numeric vector containing moving window values calculated for each grid cell
}
\description{
Calculate the mean moving window value for a given radius/shape/function for each cell in a larger resolution grid.
}
