% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlog.R
\name{xlog}
\alias{xlog}
\title{Upscale the moving window value}
\usage{
xlog(x)
}
\arguments{
\item{x}{numeric}
}
\value{
numeric vector
}
\description{
Calculate xlogx of a given value for use in `diversity()` and `diversity_nbrhd()` functions
}
