% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_torus.R
\name{create_torus}
\alias{create_torus}
\title{Pad a raster by a specified radius}
\usage{
create_torus(dat, r)
}
\arguments{
\item{dat}{The raster dataset to pad}

\item{r}{The radius by which to pad the raster}
}
\value{
raster. Original raster padded by radius r with torus effect
}
\description{
This function pads a raster by a specified number of cells, creating the effect of a
torus. Allows for moving window analysis that avoids edge effects (e.g. on simulated
landscapes).
}
\examples{
data(cat_ls)
d = create_torus(dat = cat_ls, r = 5)

}
\keyword{raster}
\keyword{torus,}
