% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity}
\alias{diversity}
\title{Calculate Shannon evenness for a given raster and list of landcover classes}
\usage{
diversity(dat, lc_class)
}
\arguments{
\item{dat}{The raster dataset on which to calculate Shannon evenness}

\item{lc_class}{The list of land cover classes for which Shannon evenness should be calculated.}
}
\value{
numeric. Shannon evenness for the raster based on the given landcover classes
}
\description{
This function calculates a value for Shannon evenness for an input raster based on a list of landcover classes.
}
\keyword{Shannon,}
\keyword{diversity,}
\keyword{evenness,}
\keyword{focal}
