% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grainchanger-package.R
\docType{package}
\name{grainchanger-package}
\alias{grainchanger}
\alias{grainchanger-package}
\title{grainchanger: Calculate Functions Using Moving Window and Upscale}
\description{
This package acts as a wrapper for the raster package and allows users to upscale fine resolution rasters to a coarser gird using specified functions either using moving windows or not.
}
\author{
\strong{Maintainer}: Laura Graham \email{LauraJaneEGraham@gmail.com}

}
\keyword{internal}
