% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grainchanger-package.R
\docType{package}
\name{grainchanger-package}
\alias{grainchanger}
\alias{grainchanger-package}
\title{grainchanger: Moving-Window and Direct Data Aggregation}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Data aggregation via moving window or direct methods. Aggregate a 
    fine-resolution raster to a grid. The moving window method smooths the surface 
    using a specified function within a moving window of a specified size and shape 
    prior to aggregation. The direct method simply aggregates to the grid using the 
    specified function.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/grainchanger}
  \item Report bugs at \url{https://github.com/ropensci/grainchanger/issues}
}

}
\author{
\strong{Maintainer}: Laura Graham \email{LauraJaneEGraham@gmail.com} (\href{https://orcid.org/0000-0002-3611-7281}{ORCID})

Other contributors:
\itemize{
  \item Felix Eigenbrod \email{f.eigenbrod@soton.ac.uk} (Input on initial conceptual development) [contributor]
  \item Marco Sciaini \email{sciaini.marco@gmail.com} (Input on package development and structure) [contributor]
}

}
\keyword{internal}
