% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomove_agg.R
\name{nomove_agg}
\alias{nomove_agg}
\title{Upscale the function value}
\usage{
nomove_agg(g, dat, fun, ...)
}
\arguments{
\item{dat}{The raster dataset on which to calculate the moving window function}

\item{...}{further arguments passed to or from other methods}

\item{grid}{the grid across which to calculate the upscaled moving window function (Raster or SpatialPolygon)}

\item{fn}{The function to apply}
}
\value{
Raster (if input is Raster) or numeric vector (if input is SpatialPolygon) containing values calculated for each coarser cell
}
\description{
Calculate the value for a given function for each cell in a larger resolution grid.
}
