% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.R
\name{prop}
\alias{prop}
\title{Calculate Shannon evenness for a given raster and list of landcover classes}
\usage{
prop(dat, lc_class)
}
\arguments{
\item{dat}{The raster dataset on which to calculate proportion}

\item{lc_class}{The list of land cover classes for which proportion should be calculated.}
}
\value{
numeric. Proportion of specified land cover classes in the raster
}
\description{
This function calculates a value for Shannon evenness for an input raster based on a list of landcover classes.
}
\keyword{composition,}
\keyword{focal}
\keyword{proportion,}
