% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.R
\name{prop}
\alias{prop}
\title{Proportion of landscape}
\usage{
prop(dat, lc_class, na.rm = TRUE)
}
\arguments{
\item{dat}{A raster or vector}

\item{lc_class}{The list of land cover classes for which proportion should be calculated.}
}
\value{
numeric. Proportion of specified land cover classes in the raster
}
\description{
This function calculates the total proportion of the landscape covered by a given land cover type(s)
}
\keyword{composition,}
\keyword{focal}
\keyword{proportion,}
