% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winmove.R
\name{winmove}
\alias{winmove}
\title{Calculate the value of a given function for each cell, within a given neighbourhood, for a cell in a larger resolution grid.}
\usage{
winmove(dat, radius, type, fn, ...)
}
\arguments{
\item{dat}{The raster dataset on which to calculate the moving window function}

\item{radius}{The radius of the moving window}

\item{type}{The shape of the moving window}

\item{fn}{The function to apply}

\item{...}{further arguments passed to or from other methods}

\item{cell}{the row identifier of the grid cell}

\item{grid}{the grid from which to take the cell}
}
\value{
A raster with the moving window values calculated
}
\description{
This function takes as input the row identifier of a cell within a large resolution grid and calculates the moving window function for a smaller resolution raster dataset using the `winmove_nbrhd()` function.
}
