% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winmove_agg.R
\name{winmove_agg}
\alias{winmove_agg}
\title{Moving-window aggregation}
\usage{
winmove_agg(g, dat, d, type, fun, ...)
}
\arguments{
\item{g}{the grid across which to calculate the upscaled moving window function
(raster or sf object)}

\item{dat}{The raster dataset on which to calculate the moving window function}

\item{d}{numeric. If `type=circle`, the radius of the circle (in units of the CRS). If
`type=rectangle` the dimension of the rectangle (one or two numbers). If `type=Gauss` the
size of sigma, and optionally another number to determine the size of the matrix
returned (default is 3 times sigma)}

\item{type}{The shape of the moving window}

\item{fun}{The function to apply}

\item{...}{further arguments passed to or from other methods}
}
\value{
Numeric vector containing moving window values calculated for each grid cell
}
\description{
Calculate the mean moving window value for a given radius/shape/function for each cell
in a larger resolution grid.
}
